requirements.txt
# Core FastAPI dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# HTTP client
httpx>=0.25.0

# Database
sqlalchemy>=2.0.0
psycopg[binary]>=3.2.0  # PostgreSQL adapter (psycopg3)

# AWS S3 integration
boto3>=1.34.0

# Configuration
tomli>=2.0.0; python_version < "3.11"  # TOML parser for Python < 3.11

# URL content extraction
newspaper3k>=0.2.8
pillow>=10.0.0
textblob>=0.17.1

# Testing (optional, for development)
pytest>=7.4.0
pytest-asyncio>=0.21.0

